Partial repair for issue #23,
The error is a result of unknown task types for original syntax checker
mechanism, added in aperte, like Mulestep, Mailstep etc. One solution is
to add all new steps to Taskfactor.java, or as it is presented in this
commit, add default checker for steps that are not known to signavio.
Its a temporary solution further forks are required. The problem is that
syntax checker is not testing  special aperte steps. The best solution
is to modifie mechanism to include them.